+++ crates/playwright/src/protocol/mod.rs
@@ -64,7 +64,7 @@ pub use file_payload::{FilePayload, FilePayloadBuilder};
 pub use frame::Frame;
 pub use keyboard::Keyboard;
 pub use local_utils::LocalUtils;
-pub use locator::{AriaRole, GetByRoleOptions, Locator};
+pub use locator::{AriaRole, FilterOptions, GetByRoleOptions, Locator};
 pub use mouse::Mouse;
 pub use page::{AddStyleTagOptions, GotoOptions, Page, Response, WaitUntil};
 pub use playwright::Playwright;